+++ src/components/RequestRoomBookingForm/RequestForm.css
@@ -59,7 +59,6 @@
 
 .form-row .form-group {
   flex: 1;
-  min-width: 150px;
 }
 
 textarea {
@@ -80,6 +79,7 @@ textarea {
   border: none;
   border-radius: 4px;
   cursor: pointer;
+  width: 100%;
 }
 
 .cancel-button {
@@ -108,10 +108,5 @@ textarea {
   .form-actions {
     flex-direction: column;
     align-items: stretch;
-  }
-
-  .cancel-button,
-  .submit-button {
-    width: 100%;
   }
 }
