+++ src/components/RequestRoomBookingForm/RoomBookingDetailsForm.js
@@ -1,0 +1,19 @@
+import React, { useState } from "react";
+
+const RoomBookingDetails = ({chosenRoom})=> {
+
+    const roomId =chosenRoom.id;
+    const checkIn = chosenRoom.
+
+    return ({
+        //codigo html
+        //Keyframe com retangulo, onde na parte de cima(maior parte) tem foto da sala, embaixo nome, localizaçao da sala (campus, andar) capacidade, recursos listados lado a lado, 
+        //pegar startTime, endTime, data no requestForm
+        //parte de detalhes da reserva:
+        //dar um titulo p reserva, descrição 
+
+    })
+}
+
+
+export default RoomBookingDetails;
